Rename video page to session page
